requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
alsa-lib=1.2.3=h516909a_0
blas=1.0=mkl
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py38h06a4308_2
chardet=4.0.0=py38h578d9bd_2
cpuonly=1.0=0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
expat=2.4.2=h9c3ff4c_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.28.5=py38h497a2fe_0
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h73d1719_1008
gst-plugins-base=1.18.5=hf529b03_3
gstreamer=1.18.5=h9f60fe5_3
icu=68.2=h9c3ff4c_0
importlib-metadata=4.10.0=py38h578d9bd_0
importlib_metadata=4.10.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
jbig=2.1=h7f98852_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jsonpickle=2.0.0=pyhd8ed1ab_0
kiwisolver=1.3.2=py38h1fd1430_1
krb5=1.19.2=hcc1bbae_3
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libclang=11.1.0=default_ha53f305_1
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=h9b69904_4
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.2=h174f98d_1
libgomp=11.2.0=h1d223b6_11
libiconv=1.16=h516909a_0
libllvm11=11.1.0=hf817b99_2
libnsl=2.0.0=h7f98852_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.18=pthreads_h8fe5266_0
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.5=hd57d9b9_1
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1004
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libzlib=1.2.11=h36c2ea0_1013
lz4-c=1.9.3=h9c3ff4c_1
matplotlib-base=3.5.1=py38hf4fb855_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
more-itertools=8.12.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
music21=7.1.0=pyhd8ed1ab_1
mysql-common=8.0.27=ha770c72_3
mysql-libs=8.0.27=hfa10184_3
ncurses=6.2=h58526e2_4
ninja=1.10.2=py38hd09550d_3
nspr=4.32=h9c3ff4c_1
nss=3.74=hb5efdd6_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py38h8e6f84c_0
pip=21.3.1=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.12.3=py38h578d9bd_8
pyqt-impl=5.12.3=py38h0ffb2e6_8
pyqt5-sip=4.19.18=py38h709712a_8
pyqtchart=5.12=py38h7400c14_8
pyqtgraph=0.12.3=pyhd8ed1ab_0
pyqtwebengine=5.12.1=py38h7400c14_8
python=3.8.12=hb7a2778_2_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-rtmidi=1.4.9=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.2=py3.8_cpu_0
qt=5.12.9=hda022c4_4
readline=8.1=h46c0cb4_0
setuptools=60.5.0=py38h578d9bd_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.37.0=h9cd32fc_0
tk=8.6.11=h27826a3_1
typing_extensions=3.10.0.2=pyh06a4308_0
unicodedata2=14.0.0=py38h497a2fe_0
webcolors=1.11.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zipp=3.7.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.1=ha95c52a_0
